requirements.txt
Markdown==3.8.2
typing==3.7.4.3
PyYAML==6.0.2
